src/lib/ai/server/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
opencv-python>=4.5.0
fastapi>=0.68.0
python-multipart>=0.0.5
uvicorn>=0.15.0
numpy>=1.21.0
Pillow>=9.0.0